pyproject.toml
[tool.poetry]
name = "newrelic-py-thread-pool"
version = "0.1.0"
description = ""
authors = ["Vicente Reig Rincon de Arellano <vicente.reig@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
newrelic = "^9.6.0"
python-dotenv = "^1.0.1"
structlog = "^24.1.0"
yappi = "^1.6.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
